Cleanup Eclipse gitignore
The good practice for Eclipse project is to share relevant portable project metadata with the project. So other people importing the project into Eclipse would immediately get the right configuration and settings without difficulty.

The only case when those files are not portable are cases when they reference absolute files, which are unlikely to happen since Maven or other dependency manager wrap dependency management and prevent from referencing absolute path directly in metadata. Also, for many project that do not use a dependency manager, the dependencies are often inside the project so they use project relative path rather then absolute ones. Project relative paths are portable.

Excluding .project, .settings, .classpath... is now considered as a bad practice for Eclipse. It's been discussed in various public resources and presented as a reminder in last EclipseCons.